Move header comp.
